+++ app/src/main/java/org/gem/indo/dooit/views/main/fragments/challenge/ChallengeFragment.java
@@ -3,6 +3,7 @@ package org.gem.indo.dooit.views.main.fragments.challenge;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
 import android.util.Log;
@@ -11,29 +12,34 @@ import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.FrameLayout;
+import android.widget.Toast;
 
 import org.gem.indo.dooit.DooitApplication;
 import org.gem.indo.dooit.R;
+import org.gem.indo.dooit.api.DooitAPIError;
+import org.gem.indo.dooit.api.DooitErrorHandler;
 import org.gem.indo.dooit.api.managers.ChallengeManager;
 import org.gem.indo.dooit.helpers.Persisted;
 import org.gem.indo.dooit.helpers.SquiggleBackgroundHelper;
 import org.gem.indo.dooit.models.challenge.BaseChallenge;
 import org.gem.indo.dooit.views.main.fragments.MainFragment;
 import org.gem.indo.dooit.views.main.fragments.challenge.fragments.ChallengeFreeformFragment;
-import org.gem.indo.dooit.views.main.fragments.challenge.fragments.ChallengeNoneFragment;
 import org.gem.indo.dooit.views.main.fragments.challenge.fragments.ChallengePictureFragment;
 import org.gem.indo.dooit.views.main.fragments.challenge.fragments.ChallengeQuizFragment;
 import org.gem.indo.dooit.views.main.fragments.challenge.fragments.ChallengeRegisterFragment;
+
+import java.net.ConnectException;
+import java.net.UnknownHostException;
 
 import javax.inject.Inject;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
-import butterknife.OnClick;
 import butterknife.Unbinder;
+import retrofit2.adapter.rxjava.HttpException;
 import rx.Subscription;
+import rx.functions.Action1;
 
 public class ChallengeFragment extends MainFragment {
 
@@ -52,8 +58,8 @@ public class ChallengeFragment extends MainFragment {
     @Inject
     Persisted persisted;
 
-    @BindView(R.id.fragment_challenge_register_button)
-    Button registerButton;
+    @BindView(R.id.fragment_challenge_container)
+    FrameLayout container;
 
     BaseChallenge challenge;
 
@@ -106,36 +112,6 @@ public class ChallengeFragment extends MainFragment {
                 return null;
         }
     }
-
-    @OnClick(R.id.fragment_challenge_register_button)
-    public void loadActivity(){
-        if(getActivity() != null)
-            ChallengeActivity.Builder.create(getActivity()).startActivity();
-    }
-
-
-    /****************
-     * Load helpers *
-     ****************/
-/*
-    private void loadTypeFragment(BaseChallenge challenge, boolean hasActive) {
-        if (challenge != null) {
-            ChallengeFragment.this.challenge = challenge;
-
-            if (getActivity() != null) {
-                FragmentTransaction ft = getChildFragmentManager().beginTransaction();
-                Fragment fragment = ChallengeRegisterFragment.newInstance(challenge, hasActive);
-                ft.replace(R.id.fragment_challenge_container, fragment, "fragment_challenge");
-                ft.commit();
-            }
-        } else {
-            FragmentTransaction ft = getChildFragmentManager().beginTransaction();
-            Fragment fragment = ChallengeNoneFragment.newInstance();
-            ft.replace(R.id.fragment_challenge_container, fragment, "fragment_challenge");
-            ft.commit();
-        }
-    }
-*/
     /************************
      * Life-cycle overrides *
      ************************/
@@ -151,20 +127,17 @@ public class ChallengeFragment extends MainFragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_challenge_register, container, false);
+        View view = inflater.inflate(R.layout.fragment_challenge, container, false);
         unbinder = ButterKnife.bind(this, view);
         SquiggleBackgroundHelper.setBackground(getContext(), R.color.grey_back, R.color.grey_fore, container);
-        //if (getActivity() != null) ChallengeActivity.Builder.create(getActivity()).startActivity();
         return view;
     }
 
-   /* @Override
+    @Override
     public void onStart() {
         super.onStart();
-        if (getActivity() != null && !childFragmentExists()) {
-
-        }
-    }*/
+            loadChallenge();
+    }
 
     @Override
     public void onStop() {
@@ -184,10 +157,9 @@ public class ChallengeFragment extends MainFragment {
     @Override
     public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
         super.onCreateOptionsMenu(menu, inflater);
-        getActivity().getMenuInflater().inflate(org.gem.indo.dooit.R.menu.menu_main, menu);
-        getActivity().getMenuInflater().inflate(org.gem.indo.dooit.R.menu.menu_main_challenge, menu);
+        getActivity().getMenuInflater().inflate(R.menu.menu_main, menu);
+        getActivity().getMenuInflater().inflate(R.menu.menu_main_challenge, menu);
     }
-
 
     /*************************
      * State-keeping methods *
@@ -223,7 +195,6 @@ public class ChallengeFragment extends MainFragment {
                 if (f == null) {
                     f = createEmptyChildFragment(page);
                     if (f == null) {
-//                        ChallengeActivity.Builder.createIntent()
                     } else {
                         f.setArguments(savedInstanceState);
                         ft.replace(R.id.fragment_challenge_container, f, "fragment_challenge");
@@ -234,6 +205,100 @@ public class ChallengeFragment extends MainFragment {
                     ft.commit();
                 }
             }
+        }
+    }
+
+    private void loadTypeFragment(BaseChallenge challenge, boolean hasActive) {
+        if (challenge != null) {
+            this.challenge = challenge;
+
+            FragmentTransaction ft = getChildFragmentManager().beginTransaction();
+            Fragment fragment = ChallengeRegisterFragment.newInstance(challenge, hasActive);
+            ft.replace(R.id.fragment_challenge_container, fragment, "fragment_challenge");
+            ft.commit();
+        }
+    }
+
+    public void loadChallenge() {
+        if (persisted.hasCurrentChallenge()) {
+            try {
+                challenge = persisted.getCurrentChallenge();
+                if (challenge.getDeactivationDate().isBeforeNow()) {
+                    //persisted challenge has expired
+                    persisted.clearCurrentChallenge();
+                    Toast.makeText(getContext(), R.string.challenge_persisted_challenge_thrown_out, Toast.LENGTH_SHORT).show();
+                    challengeSubscription = challengeManager.retrieveCurrentChallenge(false, new DooitErrorHandler() {
+                        @Override
+                        public void onError(final DooitAPIError error) {
+                            getActivity().runOnUiThread(new Runnable() {
+                                @Override
+                                public void run() {
+                                    if(error.getCause() instanceof ConnectException ||
+                                            error.getCause() instanceof UnknownHostException){
+                                        Toast.makeText(context, R.string.challenge_could_not_connect_to_server, Toast.LENGTH_SHORT).show();
+                                    }else if(error.getCause() instanceof HttpException && (((HttpException) error.getCause()).code()) == 404){
+                                        //This means no challenge could be found on the server, for now just do nothing
+                                    }
+                                }
+                            });
+                        }
+                    }).subscribe(new Action1<BaseChallenge>() {
+                        @Override
+                        public void call(BaseChallenge challenge) {
+                            loadTypeFragment(challenge, false);
+                        }
+                    });
+                } else {
+                    loadTypeFragment(challenge, true);
+                }
+            } catch (Exception e) {
+                Log.d(TAG, "Could not load persisted challenge");
+                persisted.clearCurrentChallenge();
+                Snackbar.make(container, R.string.challenge_persisted_challenge_thrown_out, Snackbar.LENGTH_LONG);
+                challengeSubscription = challengeManager.retrieveCurrentChallenge(false, new DooitErrorHandler() {
+                    @Override
+                    public void onError(final DooitAPIError error) {
+                        getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                if(error.getCause() instanceof ConnectException ||
+                                        error.getCause() instanceof UnknownHostException){
+                                    Toast.makeText(context, R.string.challenge_could_not_connect_to_server, Toast.LENGTH_SHORT).show();
+                                }else if(error.getCause() instanceof HttpException && (((HttpException) error.getCause()).code()) == 404){
+                                    //This means no challenge could be found on the server, for now just do nothing
+                                }
+                            }
+                        });
+                    }
+                }).subscribe(new Action1<BaseChallenge>() {
+                    @Override
+                    public void call(BaseChallenge challenge) {
+                        loadTypeFragment(challenge, false);
+                    }
+                });
+            }
+        }else{
+            challengeSubscription = challengeManager.retrieveCurrentChallenge(false, new DooitErrorHandler() {
+                @Override
+                public void onError(final DooitAPIError error) {
+                    getActivity().runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            if(error.getCause() instanceof ConnectException ||
+                                    error.getCause() instanceof UnknownHostException){
+                                Toast.makeText(context, R.string.challenge_could_not_connect_to_server, Toast.LENGTH_SHORT).show();
+                            }else if(error.getCause() instanceof HttpException && (((HttpException) error.getCause()).code()) == 404){
+                                //This means no challenge could be found on the server, for now just do nothing
+                            }
+                        }
+                    });
+                }
+            }).subscribe(new Action1<BaseChallenge>() {
+                @Override
+                public void call(BaseChallenge challenge) {
+                    loadTypeFragment(challenge, false);
+                }
+            });
         }
     }
